# backend/generated/src/main/java/hr/fer/openapidemo/controller/ProductsApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.20.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package hr.fer.openapidemo.controller;

import hr.fer.openapidemo.model.CreateProductRequest;
import hr.fer.openapidemo.model.ErrorResponse;
import org.springframework.lang.Nullable;
import hr.fer.openapidemo.model.ProductPage;
import hr.fer.openapidemo.model.ProductResponse;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", date = "2026-03-21T18:23:10.966472300+01:00[Europe/Zagreb]", comments = "Generator version: 7.20.0")
@Validated
@Tag(name = "Products", description = "Upravljanje proizvodima")
public interface ProductsApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CREATE_PRODUCT = "/products";
    /**
     * POST /products : Dodaj novi proizvod
     *
     * @param createProductRequest  (required)
     * @return Proizvod kreiran (status code 201)
     *         or Neispravan zahtjev — pogreška u validaciji podataka (status code 400)
     *         or Korisnik nije autenticiran ili je token istekao (status code 401)
     */
    @Operation(
        operationId = "createProduct",
        summary = "Dodaj novi proizvod",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "201", description = "Proizvod kreiran", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Neispravan zahtjev — pogreška u validaciji podataka", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Korisnik nije autenticiran ili je token istekao", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ProductsApi.PATH_CREATE_PRODUCT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ProductResponse> createProduct(
        @Parameter(name = "CreateProductRequest", description = "", required = true) @Valid @RequestBody CreateProductRequest createProductRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2026-03-10T12:00:00Z\", \"price\" : 8.0, \"name\" : \"Maskica za mobitel\", \"description\" : \"Maskica za mobitel iPhone 13\", \"id\" : 1, \"stock\" : 20, \"updatedAt\" : \"2026-03-10T12:00:00Z\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"/api/users/5\", \"error\" : \"Not Found\", \"message\" : \"Korisnik s ID-om 5 nije pronađen\", \"timestamp\" : \"2026-03-18T10:00:00Z\", \"status\" : 404 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"/api/users/5\", \"error\" : \"Not Found\", \"message\" : \"Korisnik s ID-om 5 nije pronađen\", \"timestamp\" : \"2026-03-18T10:00:00Z\", \"status\" : 404 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_DELETE_PRODUCT = "/products/{id}";
    /**
     * DELETE /products/{id} : Obriši proizvod
     *
     * @param id Jedinstveni identifikator resursa (required)
     * @return Proizvod obrisan (status code 204)
     *         or Korisnik nije autenticiran ili je token istekao (status code 401)
     *         or Traženi resurs nije pronađen (status code 404)
     */
    @Operation(
        operationId = "deleteProduct",
        summary = "Obriši proizvod",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Proizvod obrisan"),
            @ApiResponse(responseCode = "401", description = "Korisnik nije autenticiran ili je token istekao", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Traženi resurs nije pronađen", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.DELETE,
        value = ProductsApi.PATH_DELETE_PRODUCT,
        produces = { "application/json" }
    )
    default ResponseEntity<Void> deleteProduct(
        @NotNull @Min(value = 1L) @Parameter(name = "id", description = "Jedinstveni identifikator resursa", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"/api/users/5\", \"error\" : \"Not Found\", \"message\" : \"Korisnik s ID-om 5 nije pronađen\", \"timestamp\" : \"2026-03-18T10:00:00Z\", \"status\" : 404 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"/api/users/5\", \"error\" : \"Not Found\", \"message\" : \"Korisnik s ID-om 5 nije pronađen\", \"timestamp\" : \"2026-03-18T10:00:00Z\", \"status\" : 404 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_PRODUCT_BY_ID = "/products/{id}";
    /**
     * GET /products/{id} : Dohvati proizvod po ID-u
     *
     * @param id Jedinstveni identifikator resursa (required)
     * @return Proizvod pronađen (status code 200)
     *         or Traženi resurs nije pronađen (status code 404)
     */
    @Operation(
        operationId = "getProductById",
        summary = "Dohvati proizvod po ID-u",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Proizvod pronađen", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Traženi resurs nije pronađen", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductsApi.PATH_GET_PRODUCT_BY_ID,
        produces = { "application/json" }
    )
    default ResponseEntity<ProductResponse> getProductById(
        @NotNull @Min(value = 1L) @Parameter(name = "id", description = "Jedinstveni identifikator resursa", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2026-03-10T12:00:00Z\", \"price\" : 8.0, \"name\" : \"Maskica za mobitel\", \"description\" : \"Maskica za mobitel iPhone 13\", \"id\" : 1, \"stock\" : 20, \"updatedAt\" : \"2026-03-10T12:00:00Z\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"/api/users/5\", \"error\" : \"Not Found\", \"message\" : \"Korisnik s ID-om 5 nije pronađen\", \"timestamp\" : \"2026-03-18T10:00:00Z\", \"status\" : 404 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_GET_PRODUCTS = "/products";
    /**
     * GET /products : Dohvati sve proizvode
     *
     * @param page Broj stranice (počinje od 0) (optional, default to 0)
     * @param size Broj elemenata po stranici (optional, default to 20)
     * @param sort Polje i smjer sortiranja (npr. createdAt,desc) (optional)
     * @param minPrice Minimalna cijena filtra (optional)
     * @param maxPrice Maksimalna cijena filtra (optional)
     * @return Straničena lista proizvoda (status code 200)
     */
    @Operation(
        operationId = "getProducts",
        summary = "Dohvati sve proizvode",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Straničena lista proizvoda", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductPage.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = ProductsApi.PATH_GET_PRODUCTS,
        produces = { "application/json" }
    )
    default ResponseEntity<ProductPage> getProducts(
        @Min(value = 0) @Parameter(name = "page", description = "Broj stranice (počinje od 0)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "page", required = false, defaultValue = "0") Integer page,
        @Min(value = 1) @Max(value = 100) @Parameter(name = "size", description = "Broj elemenata po stranici", in = ParameterIn.QUERY) @Valid @RequestParam(value = "size", required = false, defaultValue = "20") Integer size,
        @Parameter(name = "sort", description = "Polje i smjer sortiranja (npr. createdAt,desc)", in = ParameterIn.QUERY) @Valid @RequestParam(value = "sort", required = false) @Nullable String sort,
        @DecimalMin(value = "0") @Parameter(name = "minPrice", description = "Minimalna cijena filtra", in = ParameterIn.QUERY) @Valid @RequestParam(value = "minPrice", required = false) @Nullable Double minPrice,
        @DecimalMin(value = "0") @Parameter(name = "maxPrice", description = "Maksimalna cijena filtra", in = ParameterIn.QUERY) @Valid @RequestParam(value = "maxPrice", required = false) @Nullable Double maxPrice
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"page\" : { \"number\" : 0, \"size\" : 20, \"totalPages\" : 3, \"totalElements\" : 42 }, \"content\" : [ { \"createdAt\" : \"2026-03-10T12:00:00Z\", \"price\" : 8.0, \"name\" : \"Maskica za mobitel\", \"description\" : \"Maskica za mobitel iPhone 13\", \"id\" : 1, \"stock\" : 20, \"updatedAt\" : \"2026-03-10T12:00:00Z\" }, { \"createdAt\" : \"2026-03-10T12:00:00Z\", \"price\" : 8.0, \"name\" : \"Maskica za mobitel\", \"description\" : \"Maskica za mobitel iPhone 13\", \"id\" : 1, \"stock\" : 20, \"updatedAt\" : \"2026-03-10T12:00:00Z\" } ] }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_UPDATE_PRODUCT = "/products/{id}";
    /**
     * PUT /products/{id} : Ažuriraj proizvod
     *
     * @param id Jedinstveni identifikator resursa (required)
     * @param createProductRequest  (required)
     * @return Proizvod ažuriran (status code 200)
     *         or Neispravan zahtjev — pogreška u validaciji podataka (status code 400)
     *         or Korisnik nije autenticiran ili je token istekao (status code 401)
     *         or Traženi resurs nije pronađen (status code 404)
     */
    @Operation(
        operationId = "updateProduct",
        summary = "Ažuriraj proizvod",
        tags = { "Products" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Proizvod ažuriran", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ProductResponse.class))
            }),
            @ApiResponse(responseCode = "400", description = "Neispravan zahtjev — pogreška u validaciji podataka", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "401", description = "Korisnik nije autenticiran ili je token istekao", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            }),
            @ApiResponse(responseCode = "404", description = "Traženi resurs nije pronađen", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ErrorResponse.class))
            })
        },
        security = {
            @SecurityRequirement(name = "bearerAuth")
        }
    )
    @RequestMapping(
        method = RequestMethod.PUT,
        value = ProductsApi.PATH_UPDATE_PRODUCT,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ProductResponse> updateProduct(
        @NotNull @Min(value = 1L) @Parameter(name = "id", description = "Jedinstveni identifikator resursa", required = true, in = ParameterIn.PATH) @PathVariable("id") Long id,
        @Parameter(name = "CreateProductRequest", description = "", required = true) @Valid @RequestBody CreateProductRequest createProductRequest
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"createdAt\" : \"2026-03-10T12:00:00Z\", \"price\" : 8.0, \"name\" : \"Maskica za mobitel\", \"description\" : \"Maskica za mobitel iPhone 13\", \"id\" : 1, \"stock\" : 20, \"updatedAt\" : \"2026-03-10T12:00:00Z\" }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"/api/users/5\", \"error\" : \"Not Found\", \"message\" : \"Korisnik s ID-om 5 nije pronađen\", \"timestamp\" : \"2026-03-18T10:00:00Z\", \"status\" : 404 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"/api/users/5\", \"error\" : \"Not Found\", \"message\" : \"Korisnik s ID-om 5 nije pronađen\", \"timestamp\" : \"2026-03-18T10:00:00Z\", \"status\" : 404 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"path\" : \"/api/users/5\", \"error\" : \"Not Found\", \"message\" : \"Korisnik s ID-om 5 nije pronađen\", \"timestamp\" : \"2026-03-18T10:00:00Z\", \"status\" : 404 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}