# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Web Shop API
  description: API za upravljanje korisnicima, proizvodima i narudžbama
  version: 2.0.0
  contact:
    name: Web Shop API Support
    email: support@webshop.hr
servers:
  - url: http://localhost:8080
    description: Lokalni razvojni poslužitelj
...
paths:
...
  /products:
    get:
      operationId: getProducts
      summary: Dohvati sve proizvode
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/PageParam'
        - $ref: '#/components/parameters/SizeParam'
        - $ref: '#/components/parameters/SortParam'
        - name: minPrice
          in: query
          description: Minimalna cijena filtra
          schema:
            type: number
            format: double
            minimum: 0
        - name: maxPrice
          in: query
          description: Maksimalna cijena filtra
          schema:
            type: number
            format: double
            minimum: 0
      responses:
        '200':
          description: Straničena lista proizvoda
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductPage'

    post:
      operationId: createProduct
      summary: Dodaj novi proizvod
      tags:
        - Products
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '201':
          description: Proizvod kreiran
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /products/{id}:
    get:
      operationId: getProductById
      summary: Dohvati proizvod po ID-u
      tags:
        - Products
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '200':
          description: Proizvod pronađen
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '404':
          $ref: '#/components/responses/NotFound'

    put:
      operationId: updateProduct
      summary: Ažuriraj proizvod
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateProductRequest'
      responses:
        '200':
          description: Proizvod ažuriran
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProductResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

    delete:
      operationId: deleteProduct
      summary: Obriši proizvod
      tags:
        - Products
      security:
        - bearerAuth: []
      parameters:
        - $ref: '#/components/parameters/IdParam'
      responses:
        '204':
          description: Proizvod obrisan
        '401':
          $ref: '#/components/responses/Unauthorized'
        '404':
          $ref: '#/components/responses/NotFound'

  # ─── ORDERS ──────────────────────────────────────────────────────────────────
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token dobiven putem /auth/login endpointa

  # ── Reusable parameters ────────────────────────────────────────────────────
  parameters:
    IdParam:
      name: id
      in: path
      required: true
      description: Jedinstveni identifikator resursa
      schema:
        type: integer
        format: int64
        minimum: 1
    PageParam:
      name: page
      in: query
      description: Broj stranice (počinje od 0)
      schema:
        type: integer
        default: 0
        minimum: 0
    SizeParam:
      name: size
      in: query
      description: Broj elemenata po stranici
      schema:
        type: integer
        default: 20
        minimum: 1
        maximum: 100
    SortParam:
      name: sort
      in: query
      description: "Polje i smjer sortiranja (npr. createdAt,desc)"
      schema:
        type: string
        example: createdAt,desc

  # ── Reusable responses ─────────────────────────────────────────────────────
  responses:
    BadRequest:
      description: Neispravan zahtjev — pogreška u validaciji podataka
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Unauthorized:
      description: Korisnik nije autenticiran ili je token istekao
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    NotFound:
      description: Traženi resurs nije pronađen
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'

  # ── Schemas ────────────────────────────────────────────────────────────────
  schemas:
...
    CreateProductRequest:
      type: object
      required:
        - name
        - price
        - stock
      properties:
        name:
          type: string
          minLength: 1
          maxLength: 200
          example: Bežični miš
        description:
          type: string
          example: Bežični miš s dongle USB prijemnikom
        price:
          type: number
          format: double
          minimum: 0
          exclusiveMinimum: true
          example: 10.00
        stock:
          type: integer
          minimum: 0
          example: 10
    ProductResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 1
        name:
          type: string
          example: Maskica za mobitel
        description:
          type: string
          example: Maskica za mobitel iPhone 13
        price:
          type: number
          format: double
          example: 8.00
        stock:
          type: integer
          example: 20
        createdAt:
          type: string
          format: date-time
          example: "2026-03-10T12:00:00Z"
        updatedAt:
          type: string
          format: date-time
          example: "2026-03-10T12:00:00Z"
    ProductPage:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ProductResponse'
        page:
          $ref: '#/components/schemas/PageMetadata'

    # -- Order --
...
    PageMetadata:
      type: object
      properties:
        number:
          type: integer
          description: Trenutna stranica (od 0)
          example: 0
        size:
          type: integer
          description: Broj elemenata po stranici
          example: 20
        totalElements:
          type: integer
          format: int64
          description: Ukupan broj elemenata
          example: 42
        totalPages:
          type: integer
          description: Ukupan broj stranica
          example: 3
    ErrorResponse:
      type: object
      properties:
        timestamp:
          type: string
          format: date-time
          example: "2026-03-18T10:00:00Z"
        status:
          type: integer
          example: 404
        error:
          type: string
          example: Not Found
        message:
          type: string
          example: Korisnik s ID-om 5 nije pronađen
        path:
          type: string
          example: /api/users/5